requirements.txt
Django==1.8.4
Fabric==1.10.1
SwampDragon==0.4.2.2
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
boto==2.38.0
certifi==2015.04.28
dj-database-url==0.3.0
dj-static==0.0.6
django-ipware==0.1.0
django-jsonfield==0.9.13
django-redis-cache==1.6.0
django-storages==1.1.8
django-toolbelt==0.0.1
djangorestframework==3.0.5
drf-nested-routers==0.9.0
ecdsa==0.13
eventlog==0.11.0
gunicorn==19.3.0
meld3==1.0.2
paramiko==1.15.2
psycopg2==2.6
pycrypto==2.6.1
python-dateutil==2.4.2
pytz==2015.2
redis==2.10.3
requests==2.3.0
six==1.9.0
sockjs-tornado==1.0.1
static3==0.5.1
tornado==4.2.1
tornado-redis==2.4.18
wsgiref==0.1.2
